+++ packages/snap-preact-components/src/components/Atoms/Merchandising/InlineBanner/InlineBanner.test.tsx
@@ -15,6 +15,16 @@ mockData.updateConfig({ search: 'inlineBanners.page1' });
 const searchResponse: SearchResponseModel = mockData.search();
 
 describe('Merchandising Inline Banner Component', () => {
+	beforeEach(() => {
+		// IntersectionObserver isn't available in test environment
+		const mockIntersectionObserver = jest.fn();
+		mockIntersectionObserver.mockReturnValue({
+			observe: () => null,
+			unobserve: () => null,
+			disconnect: () => null,
+		});
+		window.IntersectionObserver = mockIntersectionObserver;
+	});
 	it('renders type:inline banner', () => {
 		const rendered = render(<InlineBanner banner={searchResponse.merchandising?.content?.inline![0] as Banner} />);
 		const merchBannerElement = rendered.container.querySelector('.ss__inline-banner');