+++ packages/snap-store-mobx/src/Autocomplete/Stores/AutocompleteTermStore.ts
@@ -2,7 +2,12 @@ import { observable, makeObservable } from 'mobx';
 import type { UrlManager } from '@searchspring/snap-url-manager';
 import type { AutocompleteStateStore } from './AutocompleteStateStore';
 import type { AutocompleteStoreConfig, StoreServices } from '../../types';
-import type { AutocompleteResponseModelAllOfAutocomplete, SearchResponseModelPagination, SearchResponseModelSearch } from '@searchspring/snapi-types';
+import type {
+	AutocompleteRequestModelSearchSourceEnum,
+	AutocompleteResponseModelAllOfAutocomplete,
+	SearchResponseModelPagination,
+	SearchResponseModelSearch,
+} from '@searchspring/snapi-types';
 
 export class AutocompleteTermStore extends Array<Term> {
 	static get [Symbol.species](): ArrayConstructor {
@@ -52,7 +57,8 @@ export class AutocompleteTermStore extends Array<Term> {
 					},
 					terms,
 					resetTerms,
-					rootState
+					rootState,
+					'suggested' as AutocompleteRequestModelSearchSourceEnum
 				)
 			)
 		);
@@ -66,16 +72,19 @@ export class Term {
 	public value: string;
 	public preview: () => void;
 	public url: UrlManager;
+	public type: AutocompleteRequestModelSearchSourceEnum;
 
 	constructor(
 		services: StoreServices,
 		term: { active: boolean; value: string },
 		terms: Term[],
 		resetTerms: () => void,
-		rootState: AutocompleteStateStore
+		rootState: AutocompleteStateStore,
+		type: AutocompleteRequestModelSearchSourceEnum
 	) {
 		this.active = term.active;
 		this.value = term.value;
+		this.type = type;
 
 		this.url = services?.urlManager?.set({ query: this.value });
 
@@ -84,7 +93,7 @@ export class Term {
 			terms.map((term) => {
 				term.active = false;
 			});
-
+			rootState.source = type;
 			this.active = true;
 			rootState.locks.terms.lock();
 			rootState.locks.facets.unlock();