+++ packages/snap-store-mobx/src/Search/Stores/SearchMerchandisingStore.test.ts
@@ -10,61 +10,91 @@ const services = {
 const mockData = new MockData();
 
 describe('Merchandising Store', () => {
-	const emptyMerchStore = { redirect: '', personalized: false, experiments: [], content: {}, campaigns: [] };
+	const emptyMerchStore = { redirect: '', personalized: false, experiments: [], content: {}, campaigns: [], responseId: '' };
 
 	it('is empty when not passed valid params', () => {
 		// @ts-ignore
-		let merchStore = new SearchMerchandisingStore(services, undefined);
+		let merchStore = new SearchMerchandisingStore(services, undefined, undefined);
 		expect(merchStore).toEqual(emptyMerchStore);
 
 		// @ts-ignore
-		merchStore = new SearchMerchandisingStore(services, undefined);
+		merchStore = new SearchMerchandisingStore(services, undefined, { responseId: '' });
 		expect(merchStore).toEqual(emptyMerchStore);
 
 		// @ts-ignore
-		merchStore = new SearchMerchandisingStore(undefined, {});
-		expect(merchStore).toEqual(emptyMerchStore);
+		merchStore = new SearchMerchandisingStore(undefined, {}, undefined);
+		expect(merchStore).toEqual({ ...emptyMerchStore, responseId: '' });
 
-		merchStore = new SearchMerchandisingStore(services, {});
-		expect(merchStore).toEqual(emptyMerchStore);
+		merchStore = new SearchMerchandisingStore(services, {}, { responseId: '' });
+		expect(merchStore).toEqual({ ...emptyMerchStore, responseId: '' });
 	});
 
 	it('has banner content', () => {
 		const data = mockData.searchMeta('merchandising');
-		const merchStore = new SearchMerchandisingStore(services, data.merchandising!);
+		const merchStore = new SearchMerchandisingStore(services, data.merchandising!, data.tracking);
 
-		expect(merchStore.content).toEqual(data.merchandising?.content!);
+		const expectedContent = Object.entries(data.merchandising?.content!).reduce((acc, [type, value]) => {
+			const htmlString = value[0] || '';
+			const match = typeof htmlString === 'string' && htmlString.match(/data-banner-id="(\d+)"/);
+			const uid = match ? match[1] : '';
+
+			acc[type] = [
+				{
+					value,
+					uid,
+					responseId: data.tracking.responseId,
+				},
+			];
+			return acc;
+		}, {} as any);
+
+		expect(merchStore.content).toEqual(expectedContent);
 	});
 
 	it('has more banner content', () => {
 		const data = mockData.updateConfig({ siteId: 'ga9kq2' }).searchMeta('merchandising_page1');
-		const merchStore = new SearchMerchandisingStore(services, data.merchandising!);
+		const merchStore = new SearchMerchandisingStore(services, data.merchandising!, data.tracking);
 
-		expect(merchStore.content).toEqual(data.merchandising?.content);
+		const expectedContent = Object.entries(data.merchandising?.content!).reduce((acc, [type, value]) => {
+			const htmlString = value[0] || '';
+			const match = typeof htmlString === 'string' && htmlString.match(/data-banner-id="(\d+)"/);
+			const uid = match ? match[1] : '';
+
+			acc[type] = [
+				{
+					value,
+					uid,
+					responseId: data.tracking.responseId,
+				},
+			];
+			return acc;
+		}, {} as any);
+
+		expect(merchStore.content).toEqual(expectedContent);
 	});
 
 	it('has redirect', () => {
 		const data = mockData.resetConfig().searchMeta('redirect');
-		const merchStore = new SearchMerchandisingStore(services, data.merchandising!);
+		const merchStore = new SearchMerchandisingStore(services, data.merchandising!, data.tracking);
 		expect(merchStore.redirect).toEqual(data.merchandising?.redirect);
 	});
 
 	it('has campaigns', () => {
 		const data = mockData.updateConfig({ siteId: 'ga9kq2' }).searchMeta('merchandising_page1');
-		const merchStore = new SearchMerchandisingStore(services, data.merchandising!);
+		const merchStore = new SearchMerchandisingStore(services, data.merchandising!, data.tracking);
 		expect(merchStore.campaigns).toEqual(data.merchandising?.campaigns);
 	});
 
 	it('has personalized bool', () => {
 		const data = mockData.updateConfig({ siteId: 'ga9kq2' }).searchMeta('merchandising_page1');
-		const merchStore = new SearchMerchandisingStore(services, data.merchandising!);
+		const merchStore = new SearchMerchandisingStore(services, data.merchandising!, data.tracking);
 		expect(merchStore.personalized).toEqual(data.merchandising?.personalized);
 	});
 
 	it('updates landingPage when landing page is found', () => {
 		const data = mockData.updateConfig({ siteId: 'ga9kq2', search: 'landingPage' }).searchMeta();
 
-		const merchStore = new SearchMerchandisingStore(services, data.merchandising!);
+		const merchStore = new SearchMerchandisingStore(services, data.merchandising!, data.tracking);
 
 		expect(merchStore.campaigns).toBeDefined();
 		expect(merchStore.campaigns).toEqual(data.merchandising?.campaigns);